♻️ refactor: extract date formatting logic into a shared utility function
