+++ apps/expo/src/app/event/[id]/index.tsx
@@ -11,7 +11,6 @@ import { useSafeAreaInsets } from "react-native-safe-area-context";
 import { Image } from "expo-image";
 import { Link, router, Stack, useLocalSearchParams } from "expo-router";
 import { useUser } from "@clerk/clerk-expo";
-import { Temporal } from "@js-temporal/polyfill";
 import { EyeOff, Globe2, MapPin, User } from "lucide-react-native";
 
 import type { AddToCalendarButtonPropsRestricted } from "@soonlist/cal/types";
@@ -22,7 +21,7 @@ import LoadingSpinner from "~/components/LoadingSpinner";
 import ShareButton from "~/components/ShareButton";
 import { UserProfileFlair } from "~/components/UserProfileFlair";
 import { api } from "~/utils/api";
-import { getDateTimeInfo, timeFormatDateInfo } from "~/utils/dates";
+import { formatEventDateRange } from "~/utils/dates";
 
 export default function Page() {
   const { id } = useLocalSearchParams<{ id: string }>();
@@ -117,61 +116,11 @@ export default function Page() {
   const eventData = event.event as AddToCalendarButtonPropsRestricted;
   const isCurrentUserEvent = currentUser?.id === event.userId;
 
-  const formatDate = (date: string, startTime?: string, endTime?: string) => {
-    // Convert from event timezone to local time
-    const localTimezone = Temporal.Now.timeZoneId();
-    const eventTimezone = eventData.timeZone || localTimezone;
-    const startDateInfo = getDateTimeInfo(date, startTime || "");
-    if (!startDateInfo) return { date: "", time: "" };
-
-    // If we have a valid event timezone, convert the time
-    if (eventTimezone && eventTimezone !== "unknown") {
-      try {
-        const zonedDateTime = Temporal.ZonedDateTime.from(
-          `${date}T${startTime || "00:00:00"}[${eventTimezone}]`,
-        );
-        const localDateTime = zonedDateTime.withTimeZone(localTimezone);
-        startDateInfo.hour = localDateTime.hour;
-        startDateInfo.minute = localDateTime.minute;
-      } catch (error) {
-        console.error("Error converting timezone:", error);
-      }
-    }
-
-    const formattedDate = `${startDateInfo.dayOfWeek}, ${startDateInfo.monthName} ${startDateInfo.day}`;
-    const formattedStartTime = startTime
-      ? timeFormatDateInfo(startDateInfo)
-      : "";
-
-    let formattedEndTime = "";
-    if (endTime) {
-      const endDateInfo = getDateTimeInfo(date, endTime);
-      if (endDateInfo && eventTimezone && eventTimezone !== "unknown") {
-        try {
-          const zonedDateTime = Temporal.ZonedDateTime.from(
-            `${date}T${endTime}[${eventTimezone}]`,
-          );
-          const localDateTime = zonedDateTime.withTimeZone(localTimezone);
-          endDateInfo.hour = localDateTime.hour;
-          endDateInfo.minute = localDateTime.minute;
-        } catch (error) {
-          console.error("Error converting timezone:", error);
-        }
-      }
-      formattedEndTime = endDateInfo ? timeFormatDateInfo(endDateInfo) : "";
-    }
-
-    const timeRange =
-      startTime && endTime
-        ? `${formattedStartTime} - ${formattedEndTime}`
-        : formattedStartTime;
-    return { date: formattedDate, time: timeRange.trim() };
-  };
-
-  const { date, time } = formatDate(
+  const { date, time } = formatEventDateRange(
     eventData.startDate || "",
     eventData.startTime,
     eventData.endTime,
+    eventData.timeZone || "",
   );
 
   return (
